This is just for the vscode-speech extension
This Say.js does not work properly in the using of other TTS, like the VS Code extension vscode-speech.
The reason is that Node.js does not support Chinese encoding well.
Therefore, for Chinese user, if in need, we can add just 2-3 lines in base.js to make it work with the help of 'iconv-lite'. See in my Fork.
Of cause, the iconv-lite must be placed together in path.